% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clf_check_test.R
\name{add_new_taxa}
\alias{add_new_taxa}
\title{Create a new taxa to be considered for classification after processing with VP}
\usage{
add_new_taxa(taxa, basepath)
}
\arguments{
\item{taxa}{new taxa name to be added (can be a list of multiple taxa names)}

\item{basepath}{basepath used for clf_check}
}
\value{
empty directory structure using new taxa name inside basepath
}
\description{
creates empty directory structure to allow consideration of new taxa during clf_check()
}
