% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{add_ymd}
\alias{add_ymd}
\title{Add Year/ month/ day hour:minute:second information}
\usage{
add_ymd(data, year)
}
\arguments{
\item{data}{VPR data frame from \code{\link{ctd_roi_merge}}}

\item{year}{Year of data collection}
}
\value{
a VPR data frame with complete date/time information in a new row named 'ymd'
}
\description{
Calculate and record calendar dates for vpr data from day-of-year, hour, amd time (in milliseconds) info.
Will also add 'avg_hr' parameter if not already present.
}
