% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{bin_profile_taxa}
\alias{bin_profile_taxa}
\title{bin vpr data for a specific taxa
bins all CTD data but n_roi_bin value will be only for taxa specified}
\usage{
bin_profile_taxa(data, taxa, binSize, imageVolume)
}
\arguments{
\item{data}{a dataframe object with rows, pressure, time_ms, and number of roi data for taxa specified}

\item{taxa}{the taxa for which you want to bin data}

\item{binSize}{size of depth bins over which to average
@param imageVolume the volume of VPR images used for calculating concentrations (mm^3)}
}
\description{
bin vpr data for a specific taxa
bins all CTD data but n_roi_bin value will be only for taxa specified
}
\details{
Image volume calculations can change based on optical setting of VPR as well as autodeck setting used to process images
For IML2018051 (S2) image volume was calculated as 108155 mm^3 by seascan (6.6 cubic inches)
For COR2019002 S2 image volume was calculated as 83663 mm^3 and S3 image volume was calculated as 366082 mm^3


bins data based on KS/RK method
}
