% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{bin_vpr_data}
\alias{bin_vpr_data}
\title{Bin vpr data}
\usage{
bin_vpr_data(ctd_roi_oce, imageVolume, binSize)
}
\arguments{
\item{ctd_roi_oce}{\code{oce} ctd format VPR data from \code{\link{create.oce.vpr}}}

\item{imageVolume}{the volume of VPR images used for calculating concentrations (mm^3)}

\item{binSize}{passed to \code{\link{bin_average_vpr}}, determines size of depth bins over which data is averaged}
}
\description{
Bin vpr data
}
\details{
Image volume calculations can change based on optical setting of VPR as well as autodeck setting used to process images
For IML2018051 (S2) image volume was calculated as 108155 mm^3 by seascan (6.6 cubic inches)
For COR2019002 S2 image volume was calculated as 83663 mm^3 and S3 image volume was calculated as 366082 mm^3
}
