% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{binQC}
\alias{binQC}
\title{remove bins where less than 0.8 seconds was spent
removes spikes in binned data
based on  analysis done by R. Klaver}
\usage{
binQC(data)
}
\arguments{
\item{data}{binned data (must include 'time_diff_s' variable)}
}
\description{
remove bins where less than 0.8 seconds was spent
removes spikes in binned data
based on  analysis done by R. Klaver
}
\note{
THIS ANALYSIS MAY NO LONGER BE VALID, REQUIRES A REVISIT TO DETERMINE
  IDEAL CUT OFF VALUE. USE WITH CAUTION AS MANY VALID DATA POINTS WERE
  BEING REMOVED BASED ON THIS QC PARAMETER
}
