% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{bubble_plotTS}
\alias{bubble_plotTS}
\title{TS bubble plot for VPR data}
\usage{
bubble_plotTS(x, r)
}
\arguments{
\item{x}{oce CTD object}

\item{r}{VPR / CTD data frame containing columns salinity, temperature, n_roi}
}
\value{
TS bubble plot
}
\description{
TS bubble plot for VPR data
}
\details{
!!WARNING: This function is vulnerable to changes in oce structures and has some minimal hard coding 
with plotting details
}
\author{
R. Klaver & K. Sorochan
}
