% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{bubble_plotTSconc}
\alias{bubble_plotTSconc}
\title{Get TS bubble plot for rois - concentration}
\usage{
bubble_plotTSconc(x, r, lt, pos, inches, levels, round)
}
\arguments{
\item{x}{oce CTD object}

\item{r}{VPR data frame  object with salinity, temperature, concentration, seems to be ctd data pulled to 'concentration' object}

\item{lt}{legend title}

\item{pos}{position of legend (following legend() rules)}

\item{inches}{scales pt.cex of legend (set to 0.1)}

\item{levels}{number of legend intervals}

\item{round}{number of significant figures desired for size intervals}
}
\description{
Get TS bubble plot for rois - concentration
}
\details{
!!WARNING: Involves hard coded plot options
}
\note{
EC: (to do) Arguments could be reduced , try using ggplot with data classed
  with size, might be cleaner code (ref =
  https://r4ds.had.co.nz/data-visualisation.html , aes scaling...
  aes(size = class))
}
\author{
K. Sorochan & R. Klaver
}
