% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{bubble_plotTSnroi}
\alias{bubble_plotTSnroi}
\title{Get TS bubble plot for rois - number of rois}
\usage{
bubble_plotTSnroi(x, r)
}
\arguments{
\item{x}{oce CTD object}

\item{r}{VPR data frame ( object with salinity, temperature, n_roi )}
}
\description{
Get TS bubble plot for rois - number of rois
}
\details{
Includes hard coded plotting options and variable names, not used in final product scripts
}
\note{
EC: (to do) not used in either Rmd file (deprecated older version?)
not exported
}
\author{
K. Sorochan & R. Klaver
}
