% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{calc_roiL}
\alias{calc_roiL}
\title{calculate binned roi L-1 concentrion for a specific taxa}
\usage{
calc_roiL(data, taxa, imageVolume, binSize)
}
\arguments{
\item{data}{data frame to be run through bin_profile_taxa}

\item{taxa}{taxa of interest}

\item{imageVolume}{the volume of VPR images used for calculating concentrations (mm^3)}

\item{binSize}{the size (in metres) of the depth bin over which data will be averaged}
}
\description{
calculate binned roi L-1 concentrion for a specific taxa
}
\details{
Image volume calculations can change based on optical setting of VPR as well as autodeck setting used to process images
For IML2018051 (S2) image volume was calculated as 108155 mm^3 by seascan (6.6 cubic inches)
For COR2019002 S2 image volume was calculated as 83663 mm^3 and S3 image volume was calculated as 366082 mm^3
 #' 

Uses hard coded values for image volume calculation, 
may require updating if there are changes in sampling method
}
\author{
E. Chisholm
}
