% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{conPlot_conc}
\alias{conPlot_conc}
\title{Contour Concentration Plot}
\usage{
conPlot_conc(data, dup = "mean", bw = 1)
}
\arguments{
\item{data}{vpr depth binned data, with parameters avg_hr , pressure, and conc_m3}

\item{dup}{string defining handling of duplicates, passed to interp function ('mean', 'strip' or 'error')}

\item{bw}{bin width defining contour label intervals}
}
\description{
Makes a contour plot of ROI concentration, interpolated over time and depth
}
