% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{contour_plots}
\alias{contour_plots}
\title{contour plots for VPR and CTD data}
\usage{
contour_plots(contour_data, ctd_cont_tmp, roi_cont_tmp)
}
\arguments{
\item{contour_data}{a list of data frames which is looped from the second index, containing x, y, z columns}

\item{ctd_cont_tmp}{CTD data frame containing columns time_h and pressure}

\item{roi_cont_tmp}{ROI data from VPR, data frame containing columns time_h, pressure and n_roi}
}
\value{
filled contour VPR plot
}
\description{
This function uses oce filled.contour() to plot VPR and CTD data along with hydro data caption
}
\details{
\strong{!!!WARNING:} It is hard coded and depends on variables which are not listed in function arguments,
I was unable to find the origin of these variables. This function also contains hard coding for
plot options and sizing
}
\author{
R. Klaver & K. Sorochan
}
