% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{contour_plots_binned}
\alias{contour_plots_binned}
\title{contour plots for binned VPR and CTD data}
\usage{
contour_plots_binned(
  contour_binned_data,
  binned_cont_tmp,
  binned_cont_tmp_subset
)
}
\arguments{
\item{contour_binned_data}{a list of data frames containing x, y, z columns, this should be interpolated 
data to fill the contours of the plot}

\item{binned_cont_tmp}{data frame containing VPR data and CTD data with columns 'avg_hr' and 'pressure'
PLOTTED AS VPR PATH}

\item{binned_cont_tmp_subset}{data frame containing vPR and CTD data with columns 'avg_hr', 'pressure', and 'conc_m3',
PLOTTED AS BUBBLES}
}
\value{
filled contour VPR plot
}
\description{
contour plots for binned VPR and CTD data
}
\details{
This function uses oce filled.contour() to plot VPR and CTD data along with hydro data caption
!!!WARNING: It is hard coded and depends on variables which are not listed in function arguments, 
I was unable to find the origin of these variables. This function also contains hard coding for 
plot options and sizing
}
\examples{

 @note EC: (to do) make sure these data structures are stable add in ... for other plotting arguments
}
\author{
R. Klaver & K. Sorochan
}
