% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{ctd_cast}
\alias{ctd_cast}
\title{Isolate ascending or descending section of ctd cast}
\usage{
ctd_cast(
  data,
  cast_direction = "ascending",
  data_type,
  cutoff = 0.1,
  breaks = NULL
)
}
\arguments{
\item{data}{an \code{oce} ctd object}

\item{cast_direction}{'ascending' or 'descending' depending on desired section}

\item{data_type}{specify 'oce' or 'df' depending on class of desired output}

\item{cutoff}{Argument passed to \link[oce]{ctdFindProfiles}}

\item{breaks}{Argument passed to \link[oce]{ctdFindProfiles}}
}
\value{
Outputs either data frame or oce ctd object
}
\description{
This is an internal step required to bin data
}
\note{
\code{\link{ctdFindProfiles}} arguments for \code{minLength} and \code{cutOff} were updated to
prevent losing data (EC 2019/07/23)
}
\author{
K Sorochan, E Chisholm
}
