% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{ctd_df_cols}
\alias{ctd_df_cols}
\title{Read CTD data (SBE49) and Fluorometer data from CTD- VPR package}
\usage{
ctd_df_cols(x, col_list)
}
\arguments{
\item{x}{full filename (ctd .dat file)}

\item{col_list}{list of CTD data column names}
}
\description{
Internal use \code{\link{vpr_ctd_read}}
}
\details{
\strong{WARNING} This is hard coded to accept a specific
order of CTD data columns. The names and values in these columns can change
based on the specific instrument and should be updated before processing data
from a new VPR.

Text file format .dat file
Outputs ctd dataframe with variables time_ms, conductivity, temperature,
pressure, salinity, fluor_ref, fluorescence_mv, turbidity_ref,
turbidity_mv, altitude_NA
}
\author{
K. Sorochan, E. Chisholm
}
