% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctd_roi_oce}
\alias{ctd_roi_oce}
\title{VPR data including CTD and ROI information}
\format{
An oce package format, a 'CTD' object with VPR CTD and ROI data (1000
data rows)
}
\usage{
ctd_roi_oce
}
\description{
An oce formatted CTD object with VPR CTD and ROI data from package example
data set.
}
\keyword{datasets}
