% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{exploreImages_reclassified}
\alias{exploreImages_reclassified}
\title{Explore reclassified images}
\usage{
exploreImages_reclassified(day, hour, base_dir, taxa_of_interest, image_dir)
}
\arguments{
\item{day}{Character string, 3 digit day of interest of VPR data}

\item{hour}{Character string, 2 digit hour of interest of VPR data}

\item{base_dir}{directory path to folder containing day/hour folders in which misclassified and reclassified files are organized (eg.'C:/VPR_PROJECT/r_project_data_vis/classification files/') which would contain 'd123.h01/reclassified_krill.txt' )}

\item{taxa_of_interest}{Classification group from which to pull images}

\item{image_dir}{directory path to ROI images, eg. "E:\\\\data\\\\cruise_IML2018051\\\\", file seperator MUST BE "\\\\" in order to be recognized}
}
\value{
folders of misclassified or reclassified images inside image_dir
}
\description{
Pull image from reclassified or misclassified files produced during \code{\link{clf_check}}
}
