% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{exploreImages_taxa}
\alias{exploreImages_taxa}
\title{Explore images by depth and classification}
\usage{
exploreImages_taxa(
  data,
  min.depth,
  max.depth,
  roiFolder,
  format = "list",
  taxa_of_interest
)
}
\arguments{
\item{data}{data frame containing CTD and ROI data from \code{\link{merge_ctd_roi}}, which also contains calculated variables sigmaT and avg_hr}

\item{min.depth}{minimum depth of ROIs you are interested in looking at}

\item{max.depth}{maximum depth of ROIs you are interested in exploring}

\item{roiFolder}{directory that ROIs are within (can be very general eg. C:/data, but will be quicker to process with more specific file path)}

\item{format}{option of how images will be output, either as 'list'
a list of file names or 'image' where images will be displayed}

\item{taxa_of_interest}{character string of classification group from which to pull images}
}
\description{
Pulls images from specific depth ranges in specific classification group
}
\examples{
#determine range of interest
mid <- as.numeric(readline('Minimum depth of interest? '))
mad <- as.numeric(readline('Maximum depth of interest? '))
#run image exploration
roi_files <- exploreImages_taxa(all_dat, min.depth = mid, max.depth = mad, roiFolder = paste0('E:/data/IML2018051/rois/vpr', tow ), format = 'list', taxa = 'Calanus')

#copy image files into new directory to be browsed
roi_file_unlist <- unlist(roi_files)
newdir <- file.path(plotdir, paste0('vpr', tow, 'images_', mid, '_', mad, '_', taxa))
dir.create(newdir)
file.copy(roi_file_unlist, newdir)


}
