% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{format_oce}
\alias{format_oce}
\title{Formats list of ctd data into oce ctd objects
Input: list of ctd data frames (including -- salinity, temperature, pressure, conductivity, time_ms)
plus any other variables}
\usage{
format_oce(dat, loop)
}
\arguments{
\item{dat}{list of ctd data frames, read by read.ctdvpr.dat()}

\item{loop}{logical value indicating whther or not there are multiple hours of data to be processed (loop = TRUE if there are multiple hours)}
}
\description{
Output: oce format ctd objects (in list form)
}
\author{
E. Chisholm
}
