% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{format_size_data}
\alias{format_size_data}
\title{Format CTD and Size data from VPR}
\usage{
format_size_data(data, data_mea, taxa_of_interest, max_pressure = 100)
}
\arguments{
\item{data}{VPR dataframe from \code{\link{vpr_ctdroi_merge}}, with calculated variable sigmaT}

\item{data_mea}{VPR size data frame from \code{\link{vpr_autoid_read}}}

\item{taxa_of_interest}{a list of taxa of interest to be included in output dataframe}

\item{max_pressure}{maximum pressure cut off on data, allows comparison
between stations of different depths, should be the maximum depth which all
stations considered reach}
}
\value{
A dataframe containing VPR CTD and size data
}
\description{
Format CTD and Meas data frames into combined data frame for analysis and plotting of size data
}
