% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{get_isopycnals}
\alias{get_isopycnals}
\title{get vector to draw isopycnal lines on TS plot
Used internally to create TS plots}
\usage{
get_isopycnals(sal, pot.temp, reference.p = 0)
}
\arguments{
\item{pot.temp}{temperature vector in deg C}

\item{reference.p}{reference pressure for calculation, set to 0}
}
\description{
get vector to draw isopycnal lines on TS plot
Used internally to create TS plots
}
\author{
E. Chisholm
 
   @param sal salinity vector
}
