% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{get_roi_data}
\alias{get_roi_data}
\title{#' Get time and number of ROIs from IDs
  #'
  #'
  #' @details pulls time stamp and amount of rois from character string data
  #'   using substr(), outputs a new data frame
  #'
  #' @param x character vector with roi Id
  #'
  #'
  #'
  #'
  
  #Get time and n_rio from ROI id
  
  rois <- x$roi
  
  roi_x <- substr(rois, 1, 8)
  
  
  roi_df <- data.frame(roi = as.numeric(roi_x), taxa = x$taxa)
  roi_df$taxa <- as.character(roi_df$taxa)
  
  roi_df2 <- roi_df %>%
    dplyr::group_by(., taxa, roi) %>%
    dplyr::summarise(., n_roi = n()) %>%
    tidyr::spread(., taxa, n_roi)
  
  roi_df2 <- data.frame(roi_df2)
  roi_df2[is.na(roi_df2)] <- 0
  
  return(roi_df2)}
\usage{
get_roi_data(x)
}
\description{
#' Get time and number of ROIs from IDs
  #'
  #'
  #' @details pulls time stamp and amount of rois from character string data
  #'   using substr(), outputs a new data frame
  #'
  #' @param x character vector with roi Id
  #'
  #'
  #'
  #'
  
  #Get time and n_rio from ROI id
  
  rois <- x$roi
  
  roi_x <- substr(rois, 1, 8)
  
  
  roi_df <- data.frame(roi = as.numeric(roi_x), taxa = x$taxa)
  roi_df$taxa <- as.character(roi_df$taxa)
  
  roi_df2 <- roi_df %>%
    dplyr::group_by(., taxa, roi) %>%
    dplyr::summarise(., n_roi = n()) %>%
    tidyr::spread(., taxa, n_roi)
  
  roi_df2 <- data.frame(roi_df2)
  roi_df2[is.na(roi_df2)] <- 0
  
  return(roi_df2)
}
