% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{get_trrois_size}
\alias{get_trrois_size}
\title{Get size data from idsize files produced when training classifier
useful for getting size distribution of known rois from each taxa}
\usage{
get_trrois_size(directory, taxa, opticalSetting)
}
\arguments{
\item{directory}{cruise directory eg. 'C:/data/IML2018051/'}

\item{taxa}{list of character elements containing taxa of interest}

\item{opticalSetting}{VPR optical setting determining conversion between pixels and millimetres (options are'S0', 'S1', 'S2', or 'S3')}
}
\description{
Get size data from idsize files produced when training classifier
useful for getting size distribution of known rois from each taxa
}
