% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{getCast_EC}
\alias{getCast_EC}
\title{Isolate ascending or descending section of ctd cast
This is an internal step required to bin data 
Updated from original K. Sorochan / R. Klaver function by E. Chisholm}
\usage{
getCast_EC(
  data,
  cast_direction = "ascending",
  data_type,
  cutoff = 0.1,
  breaks = NULL
)
}
\arguments{
\item{data}{an \code{oce} ctd object}

\item{cast_direction}{'ascendcing' or 'descending' depending on desired section}

\item{data_type}{specify 'oce' or 'df' depending on class of desired output

Outputs either data frame or oce ctd object



NOTE: \code{\link{ctdFindProfiles}} arguments for \code{minLength} and \code{cutOff} were updated to
prevent losing data (EC 2019/07/23)}
}
\description{
Isolate ascending or descending section of ctd cast
This is an internal step required to bin data 
Updated from original K. Sorochan / R. Klaver function by E. Chisholm
}
