% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{getday}
\alias{getday}
\title{Get day identifier}
\usage{
getday(x)
}
\arguments{
\item{x}{A string specifying the directory and file name of the size file}
}
\value{
A string of only the day identifier (i.e., "dXXX")
}
\description{
Get day identifier
}
\author{
K Sorochan
}
