% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{gethour}
\alias{gethour}
\title{Get hour identifier}
\usage{
gethour(x)
}
\arguments{
\item{x}{A string specifying the directory and file name of the size file}
}
\value{
A string of only the hour identifier (i.e., "hXX")
}
\description{
Get hour identifier
}
\author{
K Sorochan
}
