% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{getMinShipSpeed_knots}
\alias{getMinShipSpeed_knots}
\title{Calculate minimum ship speed in knots for VPR tow}
\usage{
getMinShipSpeed_knots(frame_diff, tow_duration, x_dim)
}
\arguments{
\item{frame_diff}{numeric value for difference between frames (?)}

\item{tow_duration}{This argument is not actually used in the function}

\item{x_dim}{numeric value, might be time dimension, used in context: this value is multiplied by frames per second to calculate speed in metres/second}
}
\value{

}
\description{
Calculate minimum ship speed in knots for VPR tow
}
\details{
This function was written to calculate minimum ship speed in knots for IML2018051. 
!!WARNING: This function uses variables which are not present at function arguments, will not run on it's own. 
Also includes hard coding to convert from m/s to knots with only 3 decimal places
}
\author{
R. Klaver & K. Sorochan
}
