% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{getroiid}
\alias{getroiid}
\title{Get roi ids from string}
\usage{
getroiid(x)
}
\arguments{
\item{x}{A string specifying directory and file name of roi}
}
\value{
A string of only the 10 digit roi identifier
}
\description{
Get roi ids from string
}
\author{
K Sorochan
}
