% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{getRoiMeasurements}
\alias{getRoiMeasurements}
\title{THIS FUNCTION HAS BEEN DEPRECATED}
\usage{
getRoiMeasurements(taxafolder, nchar_folder, unit = "mm", opticalSetting)
}
\arguments{
\item{taxafolder}{path to taxa folder (base -- autoid folder)}

\item{nchar_folder}{number of characters in basepath}

\item{unit}{unit data will be output in, 'mm' (default -- millimetres) or 'px' (pixels)}

\item{opticalSetting}{VPR optical setting determining conversion between pixels and millimetres (options are 'S0', 'S1', 'S2', or 'S3')}
}
\description{
pull roi measurements from all taxa, all files
}
\note{
This function is very finicky, easily broken because it relies on character string splitting.
taxaFolder argument should not end in a backslash, please check output carefully to
ensure taxa names or ROI numbers have been properly sub string'd
}
