% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{gettaxaid}
\alias{gettaxaid}
\title{Get traxa ids from string}
\usage{
gettaxaid(x)
}
\arguments{
\item{x}{A string specifying the directory of the "taxafolder", containing the taxa id}
}
\value{
A string of only the taxa id
}
\description{
Get traxa ids from string
}
\author{
K Sorochan
}
