% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{image_check}
\alias{image_check}
\title{Remove ROI strings from aid and aidmeas files based on a manually organized folder of images}
\usage{
image_check(folder_dir, basepath)
}
\arguments{
\item{basepath}{directory path to original Visual Plankton files, specified
down to the classification group. eg.
'C:/data/cruise_IML2018051/autoid/krill'}
}
\description{
Should be used after \code{\link{image_copy}}, and manual image removal from created folders
}
\details{
@param folder_dir directory path to day hour folders containing manually
  reorganized images of a specific taxa eg.
  'C:/data/cruise_IML2018051/krill/images/' where that folder contains
  '......d123.h01/' which contains manually sorted images of krill
}
