% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{isopycnal_calculate}
\alias{isopycnal_calculate}
\title{Get vector to draw isopycnal lines on TS plot
Used internally to create TS plots}
\usage{
isopycnal_calculate(sal, pot.temp, reference.p = 0)
}
\arguments{
\item{sal}{salinity vector}

\item{pot.temp}{temperature vector in deg C}

\item{reference.p}{reference pressure for calculation, set to 0}
}
\description{
Get vector to draw isopycnal lines on TS plot
Used internally to create TS plots
}
\note{
: modified from source:\url{https://github.com/Davidatlarge/ggTS/blob/master/ggTS_DK.R}
}
\author{
E. Chisholm
}
