% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{loadrdata}
\alias{loadrdata}
\title{Load RData files}
\usage{
loadrdata(x)
}
\arguments{
\item{x}{data file names (with full path specified)}
}
\value{
a list of dataframe objects
}
\description{
Loads VPR processed RData files from station data, accounts for old naming
scheme, will load either 'data' or 'auto_measure_ls' objects into
consistently named list.
}
\note{
\strong{WARNING:} Hard coded to accept only objects named 'data' or 'auto_measure_ls'
}
\examples{

# Get filenames
path_name <- paste0(cruise, "_data_files")

names_stationdata <- list.files(path = path_name, pattern = "stationData_W58", full.names = TRUE)
names_measdata <- list.files(path = path_name, pattern = "meas_dat_W58", full.names = TRUE)

# Read in roi and measdata
stationdata <- loadrdata(names_stationdata)
measdata <- loadrdata(names_measdata)
}
\author{
K Sorochan
}
