% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{merge_ctd_roi}
\alias{merge_ctd_roi}
\title{Merge CTD and ROI data from VPR}
\usage{
merge_ctd_roi(ctd_dat_combine, roi_dat_combine)
}
\arguments{
\item{ctd_dat_combine}{a CTD dataframe from VPR processing from \code{\link{read.ctd.vpr}}}

\item{roi_dat_combine}{a data frame of roi aid data from \code{\link{read.aid}}}
}
\description{
Combines CTD data (time, hydrographic parameters), with ROI information
(identification number) into single dataframe, aligning ROI identifcation
numbers and taxa classifications with time and hydrographic parameters
}
\author{
E. Chisholm & K. SOrochan
}
