% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{myCTDplot}
\alias{myCTDplot}
\title{CTD data plot}
\usage{
myCTDplot(x, x_pos, y_pos)
}
\arguments{
\item{x}{CTD data frame}

\item{x_pos}{numeric vector containing longitude information, where first value is used to determine minimum
longitude limits on plot}

\item{y_pos}{numeric vector containing latitude information, where first value is used as minimum latitude
value on plot}
}
\value{
CTD plot
}
\description{
Returns a series of plots for summarizing ctd data,
including position, and profiles of temperature, salinity,
sigmaTheta and fluorescence
}
\details{
\strong{!!!WARNING:} this function contains hard coding for both CTD variables and lat/lon ranges
}
\author{
R. Klaver & K. Sorochan
}
