% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clf_check_test.R
\name{new_aids}
\alias{new_aids}
\title{Modifies aid and aid mea files based on manual reclassification}
\usage{
new_aids(reclassify, misclassified, basepath)
}
\arguments{
\item{reclassify}{list of reclassify files (output from clf_check())}

\item{misclassified}{list misclassify files (output from clf_check())}

\item{basepath}{base path to auto ID folder eg 'E:/autoID_EC_07032019/'}
}
\description{
Modifies aid and aid mea files based on manual reclassification
}
\author{
E. Chisholm
}
