% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{normalize}
\alias{normalize}
\title{scale matrix between 0-1
take each element of matrix dived by column total
Make sure to remove total rows before using}
\usage{
normalize(mat)
}
\arguments{
\item{mat}{a matrix to normalize}
}
\description{
scale matrix between 0-1
take each element of matrix dived by column total
Make sure to remove total rows before using
}
\note{
used internally for visualization of confusion matrices
}
