% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{normalize_matrix}
\alias{normalize_matrix}
\title{Normalize a matrix}
\usage{
normalize_matrix(mat)
}
\arguments{
\item{mat}{a matrix to normalize}
}
\description{
take each element of matrix dived by column total
}
\details{
Make sure to remove total rows before using with VP data
}
\note{
used internally for visualization of confusion matrices
}
