% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{plotTS_balloon}
\alias{plotTS_balloon}
\title{Make a balloon plot against a TS plot with ROI concentration and sort by taxa
includes isopycnal line calculations}
\usage{
plotTS_balloon(x, reference.p = 0, var)
}
\arguments{
\item{x}{dataframe with temperature, salinity, number of rois (n_roi_bin)}

\item{reference.p}{reference pressure (default at 0 for surface)- used to calculate ispycnals}

\item{var}{variable on which size of points will be based, eg conc_m3 or n_roi_bin}
}
\description{
Make a balloon plot against a TS plot with ROI concentration and sort by taxa
includes isopycnal line calculations
}
\note{
modified from source: https://github.com/Davidatlarge/ggTS/blob/master/ggTS_DK.R
}
\author{
E. Chisholm
}
