% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{plotTS_balloon_EC}
\alias{plotTS_balloon_EC}
\title{Make a balloon plot against a TS plot with ROI concentration and sort by taxa
includes isopycnal line calculations}
\usage{
plotTS_balloon_EC(x, reference.p = 0)
}
\arguments{
\item{x}{dataframe with temperature, salinity, number of rois named by taxa}

\item{reference.p}{reference pressure (default at 0 for surface)- used to calculate ispycnals}
}
\description{
WARNING HARD CODED FOR  5 TAXA, CALANUS, KRILL, ECHINODERM LARVAE, SMALL COPEPOD, CHAETOGNATHS
!! Uses isopycnal labelling method which does not label every contour
}
\note{
modified from source: https://github.com/Davidatlarge/ggTS/blob/master/ggTS_DK.R
}
