% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{profile_plot}
\alias{profile_plot}
\title{Plots VPR profiles of temperature, salinity, density, fluorescence and concentration (by classification group)}
\usage{
profile_plot(taxa_conc_n, taxa_to_plot)
}
\arguments{
\item{taxa_conc_n}{A VPR data frame with hydrographic and concentration data seperated by taxa (from \code{\link{calculate_vpr_concentrations}})}

\item{taxa_to_plot}{The specific classification groups which will be plotted}
}
\value{
A gridded object of at least 3 ggplot objects
}
\description{
This plot allows a good overview of vertical distribution of individual classification groups along with reference to hydrographic parameters.
Facet wrap is used to create distinct panels for each taxa provided
}
