% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{px_to_mm}
\alias{px_to_mm}
\title{Get conversion factor for pixels to mm for roi measurements}
\usage{
px_to_mm(x, opticalSetting)
}
\arguments{
\item{x}{an aidmea data frame (standard) to be converted into mm from pixels}

\item{opticalSetting}{the VPR setting determining the field of view and conversion factor between mm and pixels}
}
\description{
Used internally
}
\details{
converts pixels to mm using conversion factor specific to optical setting

Options for opticalSetting are 'S0', 'S1', 'S2', or 'S3'
}
