% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{read_aid}
\alias{read_aid}
\title{Read VPR aid files}
\usage{
read_aid(file_list_aid, file_list_aidmeas, export, station_of_interest)
}
\arguments{
\item{file_list_aid}{a list object of aid text files, containing roi strings. Output from matlab Visual Plankton software.}

\item{file_list_aidmeas}{a list object of aidmea text files, containing ROI measurements. Output from matlab Visual Plankton software.}

\item{export}{a character string specifying which type of data to output, either 'aid' (roi strings) or 'aidmeas' (measurement data)}
}
\description{
Read aid text files containing ROI string information or measurement data and output as a dataframe
}
\details{
Only outputs either ROI string information OR measurement data but both file types must be provided
}
\author{
E. Chisholm & K. Sorochan
}
