% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{read_aid_cnn}
\alias{read_aid_cnn}
\title{Read aid files produced by automated classification}
\usage{
read_aid_cnn(aid_file)
}
\arguments{
\item{aid_file}{a file path to an aid file produced by automated classification (with ROI path and probability value)}
}
\value{
ROI path and probability values in a table
}
\description{
Read aid files produced by automated classification
}
