% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{read_ctd_vpr}
\alias{read_ctd_vpr}
\title{Read and format CTD VPR data}
\usage{
read_ctd_vpr(ctd_files, station_of_interest)
}
\arguments{
\item{ctd_files}{full file paths to vpr ctd \code{.dat} files}

\item{station_of_interest}{VPR station name}
}
\description{
Acts as a wrapper for \code{\link{read.ctdvpr.data}}
}
\details{
Reads CTD data and adds day, hour, and station information. 
If there are multiple hours of CTD data, combines them into single dataframe.
}
\author{
E. Chisholm & K. Sorochan
}
