% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{roi_concentration_plot}
\alias{roi_concentration_plot}
\title{Concentration profile plot}
\usage{
roi_concentration_plot(x)
}
\arguments{
\item{x}{CTD / VPR data frame with column "concentration"}
}
\value{
Concentration profile plot
}
\description{
Concentration profile plot
}
\details{
Uses oce PlotProfile function to plot concentration from data frame
}
\author{
R. Klaver & K. Sorochan
}
