% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{subset.ctdvpr.data}
\alias{subset.ctdvpr.data}
\title{CTD subsetting function}
\usage{
\method{subset}{ctdvpr.data}(x)
}
\arguments{
\item{x}{CTD data frame}
}
\value{
CTD data frame with only hydrographic data
}
\description{
CTD subsetting function
}
\details{
Isolates hydrographic information from CTD dataframe output from `read.ctdvpr.data()`
Takes every 10th row of data from ctd object

!!!WARNING: This function is hard coded to row numbers inside CTD data frame, will not work with different
versions of VPR instrument or CTD package
}
\author{
R. Klaver & K. Sorochan
}
