% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{trim_ctd}
\alias{trim_ctd}
\title{subset CTD data to remove bad data at start of cast}
\usage{
trim_ctd(ctd, trimLength, end = F, endTrim = NULL)
}
\arguments{
\item{ctd}{oce CTD object}

\item{trimLength}{index of cut off point for trimming}

\item{end}{logical value (default FALSE) indicating whether to trim end of cast}

\item{endTrim}{defaults to NULL, supply number of values for trimming end of cast if end == TRUE
eg, endTrim = 5 will remove the last 5 measurmeents off the cast}
}
\description{
subset CTD data to remove bad data at start of cast
}
\details{
Use this function to remove any anomolous data collected while CTD
was soaking or on deck before cast began Not if aligning ctd and ROI data
ensure that ctd data is trimmed before aligning roi data to avoid mismatch
Note that this method only works for towed or moored ctd data where time is
continuously sampled not profile CTD data without time recording. Currently
requires you to be consious of time sampling intervals and choose a trim
length which is a multiple of that so function is able to match time
values.
}
\author{
E. Chisholm
}
