% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{trim_ctd_plot}
\alias{trim_ctd_plot}
\title{Plot a list of CTD casts to determine if they require trimming
Internally calls trim_ctd to trim casts based on user input}
\usage{
trim_ctd_plot(ctd)
}
\arguments{
\item{ctd}{a list of oce CTD objects to be trimmed}
}
\description{
Plot a list of CTD casts to determine if they require trimming
Internally calls trim_ctd to trim casts based on user input
}
\note{
If each ctd object contains multiple up/down casts,
different procedure may be required
this function only trims based on the start/end index of entire data
}
