% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vis_cm}
\alias{vis_cm}
\title{Plots normalized confusion matrix}
\usage{
vis_cm(cm, classes, type, addLabels = T, threshold = NULL)
}
\arguments{
\item{cm}{Confusion matrix (numeric)}

\item{classes}{character list of classes present in confusion matrix
(ordered)}

\item{type}{character value 'NN', 'SVM' or 'Dual', appended to 'Confusion
Matrix' to create title}

\item{addLabels}{logical value whetehr to add percentage accuracy labels to
plot (defaults to TRUE)}

\item{threshold}{numeric value which determines the minimum value of
frequency labelled on the plot on a normalized scale of 0-1 (useful for highliighting significant
disagreement)}
}
\description{
Plots normalized confusion matrix
}
\author{
E. Chisholm
}
