% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vis_unkn}
\alias{vis_unkn}
\title{Function to visualize losses to unknown category due to disagreement in Dual classifier}
\usage{
vis_unkn(cm, classes, threshold = 0, summary = T, sample_size = NULL)
}
\arguments{
\item{cm}{dual unknown confusion matrix from VP}

\item{classes}{taxa groups in order, from VP}

\item{threshold}{minimum value which will be labelled in plot}

\item{sample_size}{character string describes the sample size used to train the model being plotted (optional)
E. Chisholm May 2019}
}
\description{
Makes confusion matrix like plot, where x axis represent SVM classification, y axis represent NN classification
Allows visual summary of data lost to unknown category
}
