% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vp_plot_matrix}
\alias{vp_plot_matrix}
\title{Plots normalized confusion matrix}
\usage{
vp_plot_matrix(cm, classes, type, addLabels = TRUE, threshold = NULL)
}
\arguments{
\item{cm}{Confusion matrix (numeric)}

\item{classes}{character list of classes present in confusion matrix
(ordered)}

\item{type}{character value 'NN', 'SVM' or 'Dual', appended to 'Confusion
Matrix' to create title}

\item{addLabels}{logical value whether to add percentage accuracy labels to
plot (defaults to TRUE)}

\item{threshold}{numeric value which determines the minimum value of
frequency labelled on the plot on a normalized scale of 0-1 (useful for highlighting significant
disagreement)}
}
\value{
a visualization of the confusion matrix, normalized
}
\description{
Plots normalized confusion matrix
}
\author{
E. Chisholm
}
