% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{VPR processing functions depend on these packages

These packages are needed!}
\alias{VPR processing functions depend on these packages

These packages are needed!}
\title{Packages}
\description{
Packages
}
