% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_cnn_read}
\alias{vpr_cnn_read}
\title{Read prediction output from a CNN model}
\usage{
vpr_cnn_read(filename)
}
\arguments{
\item{filename}{model prediction output file (.txt) from \code{vpr_transferlearn::save_output()}}
}
\value{
a dataframe
}
\description{
Read prediction output from a CNN model
}
