% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_ctd_files}
\alias{vpr_ctd_files}
\title{Create a list of ctd files to be read}
\usage{
vpr_ctd_files(castdir, cruise, day_hour)
}
\arguments{
\item{castdir}{root directory for ctd cast files}

\item{cruise}{cruise name (exactly as in directory structure)}

\item{day_hour}{vector of day-hour combinations (e.g, dXXX.hXX)}
}
\value{
vector of ctd file paths matching days-hour combinations provided
}
\description{
Searches through typical VP directory structure
}
\details{
Use with caution
}
\author{
E. Chisholm and K. Sorochan
}
