% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_ctd_ymd}
\alias{vpr_ctd_ymd}
\title{Add Year/ month/ day hour:minute:second information}
\usage{
vpr_ctd_ymd(data, year, offset)
}
\arguments{
\item{data}{VPR data frame from \code{\link{vpr_ctdroi_merge}}}

\item{year}{Year of data collection}

\item{offset}{time offset in hours between VPR CPU and processed data times (optional)}
}
\value{
A VPR data frame with columns for date and time (i.e., column 'ymdhms') and hour (i.e., column time_hr)
}
\description{
Obtain columns for date and time (i.e., column "ymdhms") and time in hours (i.e., column time_hr) for each row in VPR data frame by utilizing day-of-year, hour, and millisecond outputs from VPR data output.
}
\examples{
year <- 2019
data('ctd_roi_merge')
dat <- vpr_ctd_ymd(ctd_roi_merge, year)


}
