% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_day}
\alias{vpr_day}
\title{Get day identifier}
\usage{
vpr_day(x)
}
\arguments{
\item{x}{A string specifying the directory and file name of the size file}
}
\value{
A string of only the day identifier (i.e., "dXXX")
}
\description{
Get day identifier
}
\examples{
day_string <- 'C:/data/cruise/autoid/Calanus/d000/h00'
vpr_day(day_string)

}
\seealso{
\code{\link{vpr_hour}}, \code{\link{vpr_roi}}, \code{\link{vpr_category}}
}
\author{
K Sorochan
}
