% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_dayhour}
\alias{vpr_dayhour}
\title{Find day & hour info to match each station of interest for processing}
\usage{
vpr_dayhour(stations, file)
}
\arguments{
\item{stations}{a vector of character values naming stations of interest}

\item{file}{CSV file containing 'day', 'hour', 'station', and 'day_hour' columns}
}
\value{
Vector of day-hour combinations corresponding to stations of interest
}
\description{
@author E. Chisholm and K. Sorochan
}
