% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_hour}
\alias{vpr_hour}
\title{Get hour identifier}
\usage{
vpr_hour(x)
}
\arguments{
\item{x}{A string specifying the directory and file name of the size file}
}
\value{
A string of only the hour identifier (i.e., "hXX")
}
\description{
Get hour identifier
}
\examples{
hour_string <- 'C:/data/cruise/autoid/Calanus/d000/h00'
vpr_hour(hour_string)

}
\seealso{
\code{\link{vpr_day}}, \code{\link{vpr_roi}}, \code{\link{vpr_category}}
}
\author{
K Sorochan
}
