% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_img_check}
\alias{vpr_img_check}
\title{Remove ROI strings from aid and aidmeas files based on a manually organized folder of images}
\usage{
vpr_img_check(folder_dir, basepath)
}
\arguments{
\item{folder_dir}{directory path to day hour folders containing manually
reorganized images of a specific taxa eg.
'C:/data/cruise_IML2018051/krill/images/' where that folder contains
'......d123.h01/' which contains manually sorted images of krill}

\item{basepath}{directory path to original Visual Plankton files, specified
down to the classification group. eg.
'C:/data/cruise_IML2018051/autoid/krill'}
}
\description{
Should be used after \code{\link{vpr_img_copy}}, and manual image removal from created folders
}
