% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_img_copy}
\alias{vpr_img_copy}
\title{Image copying function for specific taxa of interest}
\usage{
vpr_img_copy(auto_id_folder, taxas.of.interest, day, hour)
}
\arguments{
\item{auto_id_folder}{eg "D:/VP_data/IML2018051/autoid"}

\item{taxas.of.interest}{eg. taxas.of.interest <- c('Calanus')}

\item{day}{character, day of interest}

\item{hour}{character, hour of interest}
}
\description{
This function can be used to copy images from a particular taxa, day and hour into distinct folders within the auto id directory
This is useful for visualizing the ROIs of a particular classification group or for performing manual tertiary checks to remove
images not matching classification group descriptions.
}
