% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_img_depth}
\alias{vpr_img_depth}
\title{Explore VPR images by depth bin}
\usage{
vpr_img_depth(data, min.depth, max.depth, roiFolder, format = "list")
}
\arguments{
\item{data}{data frame containing CTD and ROI data from
\code{\link{vpr_ctdroi_merge}}, which also contains calculated variables
sigmaT and time_hr}

\item{min.depth}{minimum depth of ROIs you are interested in looking at}

\item{max.depth}{maximum depth of ROIs you are interested in exploring}

\item{roiFolder}{directory that ROIs are within (can be very general eg.
C:/data, but will be quicker to process with more specific file path)}

\item{format}{option of how images will be output, either as 'list' a list
of file names or 'image' where images will be displayed}
}
\description{
Allows user to pull VPR images from specific depth ranges, to investigate
trends before classification of images into taxa groups
}
