% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_plot_histsize}
\alias{vpr_plot_histsize}
\title{Plot size frequency histogram}
\usage{
vpr_plot_histsize(data, param, title = NULL, bw = 0.1, xlim = NULL)
}
\arguments{
\item{data}{size data from auto_measure_mm subset into taxas}

\item{param}{size parameter of interest (corresponds to sub lists within data argument)}

\item{title}{main title for plot, if left null will default based on parameter and taxa}

\item{bw}{bin width, defines width of bars on histogram, defaults to 0.1, decrease for more detail}

\item{xlim}{plot xlimit, defaults to min max of data if not provided}
}
\description{
Plot size frequency histogram
}
\note{
param options are typically 'Perimeter', 'Area', 'width1','width2',
'width3', 'short_axis_length', 'long_axis_length'
}
\author{
E. Chisholm
}
