% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_plot_sizefreq}
\alias{vpr_plot_sizefreq}
\title{Size Frequency plots for VPR data}
\usage{
vpr_plot_sizefreq(x, number_of_classes, colour_of_bar)
}
\arguments{
\item{x}{a data frame with columns 'taxa', 'long_axis_length'}

\item{number_of_classes}{numeric value passed to nclass argument in hist()}

\item{colour_of_bar}{character value defining colour of plotted bars}
}
\description{
This uses the \code{\link{hist}} plot function in base R to give a histogram of size (long axis length) frequency within a taxa.
\strong{!!WARNING:} this function uses hard coded plot attributes
}
\author{
K. Sorochan
}
