% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_plot_TS}
\alias{vpr_plot_TS}
\title{Make a balloon plot against a TS plot}
\usage{
vpr_plot_TS(x, reference.p = 0, var)
}
\arguments{
\item{x}{dataframe with temperature, salinity, number of rois (n_roi_bin)}

\item{reference.p}{reference pressure (default at 0 for surface)- used to calculate isopycnals}

\item{var}{variable on which size of points will be based, eg conc_m3 or n_roi_bin}
}
\description{
TS balloon plot with ROI concentration, sorted by taxa
includes isopycnal line calculations
}
\note{
modified from source: \url{https://github.com/Davidatlarge/ggTS/blob/master/ggTS_DK.R}
}
\author{
E. Chisholm
}
