% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_roi}
\alias{vpr_roi}
\title{Get roi ids from string}
\usage{
vpr_roi(x)
}
\arguments{
\item{x}{A string specifying directory and file name of roi}
}
\value{
A string of only the 10 digit roi identifier
}
\description{
Get roi ids from string
}
\examples{

roi_string <- 'roi.0100000000.tif'
vpr_roi(roi_string)

}
\seealso{
\code{\link{vpr_hour}}, \code{\link{vpr_day}}, \code{\link{vpr_category}}
}
\author{
K Sorochan
}
