% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_summary}
\alias{vpr_summary}
\title{Data Summary Report}
\usage{
vpr_summary(all_dat, fn, tow = tow, day = day, hour = hour)
}
\arguments{
\item{all_dat}{data frame containing VPR and CTD data including time_ms,
time_hr, conductivity, temperature, pressure, salinity, fluorescence_mv,
turbidity_mv, sigmaT}

\item{fn}{file name to save data summary, if not provided, summary will print to console}

\item{tow}{VPR tow number}

\item{day}{julian day}

\item{hour}{two digit hour (24 hr clock)}
}
\description{
Part of VP easy plot processing, prints data summary report to give quantitative, exploratory analysis of data
}
\author{
E Chisholm
}
