% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{calculate_vpr_concentrations}
\alias{calculate_vpr_concentrations}
\title{Calculate VPR concentrations}
\usage{
calculate_vpr_concentrations(data, taxas_list, station_of_interest)
}
\arguments{
\item{data}{a VPR dataframe as produced by \code{\link{merge_ctd_roi}}}

\item{taxas_list}{a list of character strings representing taxa present in the station being processed}

\item{station_of_interest}{The station being processed}
}
\description{
Calculate VPR concentrations
}
\details{
Calculates concentrations for each named taxa in dataframe
}
