% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{conPlot_EC}
\alias{conPlot_EC}
\title{Interpolated contour plot of particular variable}
\usage{
conPlot_EC(data, var, dup = "mean", method = "interp", labels = TRUE, bw = 1)
}
\arguments{
\item{data}{data frame needs to include avg_hr, pressure, and variable of
choice (var)}

\item{var}{variable in dataframe which will be interpolated and plotted}

\item{dup}{if method == 'interp'. Method of handling duplicates in interpolation, passed to interp function (options: 'mean', 'strip', 'error')}

\item{method}{Specifies interpolation method, options are 'akima', 'interp'
or 'oce', akima and interp produce indentical interpolations, oce uses
slightly different method (oce is least error prone)}

\item{labels}{logical value indicating whether or not to plot contour labels}

\item{bw}{bin width defining interval at which contours are labeled}
}
\description{
Creates interpolated contour plot, can be used as a background for ROI or tow yo information
}
\author{
E. Chisholm
}
