% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{plot_profile_conc}
\alias{plot_profile_conc}
\title{plot a profile of taxa specific ROI concetntration}
\usage{
plot_profile_conc(data, taxa, binSize, imageVolume)
}
\arguments{
\item{data}{dataframe which will be run through bin_profile_taxa (see function requirements)}

\item{taxa}{taxa of interest}

\item{binSize}{required for bin_profile_taxa (size in metres over which data will be averaged)}

\item{imageVolume}{required for bin_profile_taxa (volume of ROI images calculated in mm^3)}
}
\description{
plot a profile of taxa specific ROI concetntration
}
\note{
this function contains hard coding of concentration calculation based on image volume 
and may require updates with different sampling methods
}
\author{
E. Chisholm
}
