% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{read.ctdvpr.data}
\alias{read.ctdvpr.data}
\title{Read CTD data (SBE49) and Fluorometer data from CTD- VPR package}
\usage{
read.ctdvpr.data(x)
}
\arguments{
\item{x}{full filename (ctd .dat file)}
}
\description{
Internal use \code{\link{read_ctd_vpr}}
}
\details{
**WARNING** \code{\link{read.ctdvpr.data}} is hard coded to accept a specific
order of CTD data columns. The names and values in these columns can change
based on the specific insturment and should be updated before processing data
from a new VPR.

Text file format .dat file
Outputs ctd dataframe with variables time_ms, conductivity, temperature,
pressure, salinity, fluor_ref, fluorescence_mv, turbidity_ref,
turbidity_mv, altitude_NA
}
\author{
K. Sorochan, R. Klaver, E. Chisholm
}
