% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_autoid_check}
\alias{vpr_autoid_check}
\title{Checks manually created aid files for errors}
\usage{
vpr_autoid_check(new_autoid, original_autoid, cruise, dayhours)
}
\arguments{
\item{new_autoid}{file path to autoid folder eg. C:/data/CRUISENAME/autoid/ (produced by \code{vpr_autoid_create()})}

\item{original_autoid}{file path to original autoid folder (produced by automated classification)}

\item{cruise}{name of cruise which is being checked}

\item{dayhours}{chr vector, of unique day and hour values to check through (format d123.h12)}
}
\value{
text file (saved in working directory) named CRUISENAME_aid_file_check.txt
}
\description{
Removes any empty aid files after manual reclassification, checks for tow
numbers and other metadata to match. Performs check to ensure measurement and
ROI files are the same length
}
\details{
WARNING: This function will delete empty aid and aidmeas files, permanently changing your directory. Consider making a back up copy before running this function.
}
\author{
E Chisholm
}
