% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_autoid_check}
\alias{vpr_autoid_check}
\title{Checks manually created aid files for errors}
\usage{
vpr_autoid_check(basepath, cruise)
}
\arguments{
\item{basepath}{basepath to autoid folder eg. C:/data/CRUISENAME/autoid/}

\item{cruise}{name of cruise which is being checked}
}
\value{
text file (saved in working directory) named CRUISENAME_aid_file_check.txt
}
\description{
Removes any empty aid files after manual reclassification, checks for tow
numbers and other metadata to match. Performs check to ensure measurement and
ROI files are the same length
}
\author{
E Chisholm
}
