% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_autoid_read}
\alias{vpr_autoid_read}
\title{Read VPR aid files}
\usage{
vpr_autoid_read(
  file_list_aid,
  file_list_aidmeas,
  export,
  station_of_interest,
  opticalSetting,
  warn = TRUE
)
}
\arguments{
\item{file_list_aid}{a list object of aid text files, containing roi strings. Output from matlab Visual Plankton software.}

\item{file_list_aidmeas}{a list object of aidmea text files, containing ROI measurements. Output from matlab Visual Plankton software.}

\item{export}{a character string specifying which type of data to output, either 'aid' (roi strings) or 'aidmeas' (measurement data)}

\item{station_of_interest}{Station information to be added to ROI data output, use NA if irrelevant}

\item{opticalSetting}{Optional argument specifying VPR optical setting. If provided will be used to convert size data into mm from pixels, if missing size data will be output in pixels}

\item{warn}{Logical, FALSE silences size data unit warnings}
}
\description{
Read aid text files containing ROI string information or measurement data and output as a dataframe
}
\details{
Only outputs either ROI string information OR measurement data but both file types must be provided
}
\note{
Full paths to each file should be specified
}
\author{
E. Chisholm & K. Sorochan
}
