% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_category}
\alias{vpr_category}
\title{Get category ids from string}
\usage{
vpr_category(x)
}
\arguments{
\item{x}{A string specifying the directory of the "categoryfolder", containing the category id}
}
\value{
A string of only the category id
}
\description{
Get category ids from string
}
\examples{
category_string <- 'C:/data/cruise/autoid/Calanus/d000/h00'
vpr_category(category_string)

}
\seealso{
\code{\link{vpr_hour}}, \code{\link{vpr_day}}, \code{\link{vpr_roi}}
}
\author{
K Sorochan
}
