% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clf_check_test.R
\name{vpr_category_create}
\alias{vpr_category_create}
\title{Create a new category to be considered for classification after processing with VP}
\usage{
vpr_category_create(category, basepath)
}
\arguments{
\item{category}{new category name to be added (can be a list of multiple category names)}

\item{basepath}{path to folder containing autoid files (e.g., 'extdata/COR2019002/autoid')}
}
\value{
empty directory structure using new category name inside basepath
}
\description{
creates empty directory structure to allow consideration of new category during vpr_manual_classification()
}
