% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_ctdroisize_merge}
\alias{vpr_ctdroisize_merge}
\title{Format CTD and Size data from VPR}
\usage{
vpr_ctdroisize_merge(data, data_mea, taxa_of_interest)
}
\arguments{
\item{data}{VPR dataframe from \code{\link{vpr_ctdroi_merge}}, with calculated variable sigmaT}

\item{data_mea}{VPR size data frame from \code{\link{vpr_autoid_read}}}

\item{taxa_of_interest}{a list of taxa of interest to be included in output dataframe}
}
\value{
A dataframe containing VPR CTD and size data
}
\description{
Format CTD and Meas data frames into combined data frame for analysis and plotting of size data
}
