% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_export}
\alias{vpr_export}
\title{Format and export VPR data for publication (IN DEVELOPMENT)
Exports a csv file with standard column names based on British Oceanographic
Data Centre, BODC::P01 and DarwinCore (DwC) naming conventions,
and a JSON metadata file for station level metadata}
\usage{
vpr_export(data, metadata, columnNames)
}
\arguments{
\item{data}{a VPR data frame}

\item{metadata}{(optional) a named list of character values giving metadata to be included in JSON file}

\item{columnNames}{(optional) a named list of character values giving relationships between existing names
of data columns and standard names}
}
\description{
Format and export VPR data for publication (IN DEVELOPMENT)
Exports a csv file with standard column names based on British Oceanographic
Data Centre, BODC::P01 and DarwinCore (DwC) naming conventions,
and a JSON metadata file for station level metadata
}
