% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC_functions.R
\name{vpr_summary}
\alias{vpr_summary}
\title{Data Summary Report}
\usage{
vpr_summary(all_dat, fn, tow = tow, day = day, hour = hour)
}
\arguments{
\item{all_dat}{data frame containing VPR and CTD data including time_ms,
avg_hr, conductivity, temperature, pressure, salinity, fluorescence_mv,
turbidity_mv, sigmaT}

\item{fn}{file name to save data summary}
}
\value{

}
\description{
Part of VP easy plot processing, prints data summary report to give quantitative, exploratory analysis of data
}
\author{
E Chisholm
}
